pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[project]
name = "ngsPETSc"
version = "0.2.0"
description = "NGSolve/Netgen interface to PETSc"
authors = [
  { name = "Umberto Zerbinati", email = "umberto.zerbinati@maths.ox.ac.uk" },
  { name = "Patrick E. Farrell", email = "patrick.farrell@maths.ox.ac.uk" },
  { name = "Stefano Zampini", email = "stefano.zampini@kaust.edu.sa" },
  { name = "Jack Betteridge", email = "J.Betteridge@imperial.ac.uk" },
  { name = "Pablo D. Brubeck", email = "brubeck@protonmail.com"},
]
maintainers = [
  { name = "Umberto Zerbinati", email = "umberto.zerbinati@maths.ox.ac.uk" },
  { name = "Pablo D. Brubeck", email = "brubeck@protonmail.com"},
]
readme = "README.md"
license = "MIT"
requires-python = ">=3.10,<4.0"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Development Status :: 3 - Alpha",
]
dependencies = [
  "netgen-mesher >= 6.2.2603,<7.0",
  "petsc4py >=3.22,<4.0",
  "numpy >=2,<3",
  "scipy >=1,<2",
]

[project.urls]
Documentation = "https://ngspetsc.readthedocs.io/en/latest/"
Repository = "https://github.com/NGSolve/ngsPETSc"

[tool.poetry.extras]
firedrake = ["firedrake"]

[tool.poetry]
packages = [{ include = "ngsPETSc" }]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
pylint = "^3.3"

[tool.poetry.group.ngsolve.dependencies]
ngsolve = "^6.2"
mpi4py = "^4"
